[Refactor] #14: Building, Door 패키지 구조 변경
