+++ src/main/java/com/example/BarrierKU/domain/building/dto/BuildingResponse.java
@@ -1,9 +1,9 @@
-package com.example.BarrierKU.dto.response;
+package com.example.BarrierKU.domain.building.dto;
 
 import com.example.BarrierKU.domain.Indoor.Building;
-import com.example.BarrierKU.domain.Indoor.Door;
 import com.example.BarrierKU.domain.Indoor.Significant;
 import com.example.BarrierKU.domain.Type.Purpose;
+import com.example.BarrierKU.domain.door.dto.DoorResponse;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
